pyproject.toml
[project]
name = "pline-classifier"
version = "0.1.0"
description = "P-line ownership classification agent"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "openai>=1.60.0",
    "anthropic>=0.42.0",
    "pydantic>=2.10.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
]
